Add files via upload
This code will display the country (region) and city information if available and different. If the city information is not available or is the same as the country information, it won't be displayed.If you require city information for a wider range of phone numbers, you may need to explore other sources or APIs that provide more comprehensive geolocation data. Keep in mind that city information for all phone numbers may not be readily accessible due to privacy and data availability constraints.